Stomp.js function additions
Added stomp.js functions for login and registration
